Initial Setup + Choose libraries  + Add a menu drawer
For the UI components and styling I have chosen a robust Chakra UI library, which should be enough to create a fairly unique website. Because this website is multi-paged I will also be using React Router for client-side routing. There are also some additional small packages (and not so small) that I will be using. But the 2 core libraries are stated above